refactor: Move MyTagListResponse to dto package and update references
