+++ src/main/java/com/mycodingtest/solvedproblem/SolvedProblemController.java
@@ -2,7 +2,7 @@ package com.mycodingtest.solvedproblem;
 
 import com.mycodingtest.security.CustomUserDetails;
 import com.mycodingtest.solvedproblem.dto.SolvedProblemWithReviewResponse;
-import com.mycodingtest.tag.MyTagListResponse;
+import com.mycodingtest.tag.dto.MyTagListResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
